refactor: se deja menuSeleccion en el archivo de menu y se acomodan los menus para funcionar con cambios hechos en mecanicas
